Add files via upload
A To-Do List application made with Python, running in the console, is a straightforward and efficient way to manage tasks. Its simplicity makes it easy to use while providing essential features like task prioritization, due dates, and persistent storage. This application is a valuable tool for anyone looking to enhance their productivity without the need for a graphical interface.